requirements.txt
google-cloud-storage==2.7.0
google-auth==2.16.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.1.0
google-api-python-client==2.85.0
httplib2==0.21.0
six==1.16.0

